Launch the Shiny app

Launch the Shiny app

Usage

ovva_shiny(
  data_path,
  playlist_handler = ovva_playlist_handler(),
  highlight_handler = ovva_highlight_handler(),
  video_server = "lighttpd",
  launch_browser = TRUE,
  video_timing_df = ov_video_timing_df(),
  ...
)

Arguments

data_path

character or function: a named character vector of paths to data files. The names will be used as the competition names. So e.g. c("Competition 1" = "/path/to/dvw/files", "Competition 2" = "/path/to/other/dvw/files"). If data_path is a function, it should return such a character vector

playlist_handler

tibble: a tibble that provides playlist handler capabilities (see ovva_playlist_handler for details)

highlight_handler

tibble: a tibble that provides playlist handler capabilities (see ovva_highlight_handler for details)

video_server

string or function: if string, either "lighttpd", "servr", or "none". If a function, it will be used to modify the video file path present in each dvw file. Details TBD

launch_browser

logical: if TRUE, launch the app in the system's default web browser (passed to runApp's launch.browser parameter). If NULL, don't launch the app, just return the shinyApp object

video_timing_df

data.frame: data.frame of default clip timings

...

: additional parameters passed to the UI and server functions

See also

Examples

if (FALSE) { # \dontrun{
  ## to use your own data, assuming that your data files are in the data/volley/ folder
  ovva_shiny(data_path = c(MyData = "data/volley"))

  ## or to use the bundled demonstration file
  ovva_shiny_demo()
} # }